+++ src/main/java/com/finalproject/team4/shouldbe/controller/MypageController.java
@@ -32,6 +32,23 @@ public class MypageController {
 
     EncryptUtil encrypt = new EncryptUtil();
 
+    @GetMapping("/mypage")
+    public ModelAndView mypage(HttpSession session){
+        ModelAndView mav = new ModelAndView();
+        if(session.getAttribute("logStatus") != "Y"){
+            mav.setViewName("redirect:/login");
+            return mav;
+        }
+        String userid = (String) session.getAttribute("logId");
+        MypageVO myvo = service.mypage_info(userid);
+        System.out.println(myvo);
+
+        mav.addObject("myvo", myvo);
+        mav.setViewName("mypage/mypage");
+        return mav;
+    }
+
+
     // 마이페이지 프로필
     @GetMapping("/mypage/change_user")
     public ModelAndView mypage_change_user(HttpSession session) {
